+++ examples/nucleo_l552ze-q/adc_basic/main.cpp
@@ -44,7 +44,7 @@ main()
 		MODM_LOG_INFO << "adcValue=" << adcValue;
 		const float voltage = adcValue * 3.3 / 0xfff;
 		MODM_LOG_INFO << " voltage=";
-		MODM_LOG_INFO.printf("%.3f\n", voltage);
+		MODM_LOG_INFO.printf("%.3f\n", (double)voltage);
 		Board::Leds::toggle();
 		modm::delay(500ms);
 	}
